requirements.txt
streamlit
openai
numpy
pandas
scikit-learn
sentence-transformers
streamlit_mermaid
faiss-cpu
tiktoken
rank_bm25
st-clickable-images
unstructured
sendgrid
dotenv
orjson
tqdm
